Complete end-to-end funtionality for CNN. 3 functions are templated. 3 TOGOgit status
